requirements.txt
tqdm~=4.65.0
appdirs~=1.4.4
python-slugify~=8.0.1
discord.py-self~=2.0.0
python-dotenv~=1.0.0
httpx~=0.24.1